requirements.txt
gradio
praat-parselmouth
joblib
numpy
scikit-learn
streamlit-lottie

